refactor: canvas handlers classes
